+++ src/core/hw/tegra_x1/gpu/renderer/shader_decompiler/decoder/half_arithmetic.cpp
@@ -123,8 +123,9 @@ void EmitHadd2C(DecoderContext& context, InstHadd2C inst) {
         context, inst.base.pred, inst.base.pred_inv, inst.sat,
         inst.base.out_fmt, inst.base.dst, inst.base.src_a, inst.base.swizzle_a,
         inst.base.abs_a, inst.neg_a,
-        ir::Value::ConstMemory(CMem(inst.cbuf_slot, RZ, inst.cbuf_offset * 4),
-                               ir::VectorType(ir::ScalarType::F16, 2)),
+        ir::Value::ConstMemory(
+            CMem(inst.cbuf_slot, RZ, static_cast<u64>(inst.cbuf_offset * 4)),
+            ir::VectorType(ir::ScalarType::F16, 2)),
         inst.abs_b, inst.neg_b);
 }
 
@@ -158,8 +159,9 @@ void EmitHmul2C(DecoderContext& context, InstHmul2C inst) {
         context, inst.base.pred, inst.base.pred_inv, inst.sat,
         inst.base.out_fmt, inst.base.dst, inst.base.src_a, inst.base.swizzle_a,
         inst.base.abs_a, inst.neg_a,
-        ir::Value::ConstMemory(CMem(inst.cbuf_slot, RZ, inst.cbuf_offset * 4),
-                               ir::VectorType(ir::ScalarType::F16, 2)),
+        ir::Value::ConstMemory(
+            CMem(inst.cbuf_slot, RZ, static_cast<u64>(inst.cbuf_offset * 4)),
+            ir::VectorType(ir::ScalarType::F16, 2)),
         inst.abs_b, false);
 }
 
@@ -196,8 +198,9 @@ void EmitHfma2RC(DecoderContext& context, InstHfma2RC inst) {
         inst.base.out_fmt, inst.base.dst, inst.base.src_a, inst.base.swizzle_a,
         false, GetSwizzledHalf(context.builder, inst.swizzle_b, inst.src_b),
         inst.neg_b,
-        ir::Value::ConstMemory(CMem(inst.cbuf_slot, RZ, inst.cbuf_offset * 4),
-                               ir::VectorType(ir::ScalarType::F16, 2)),
+        ir::Value::ConstMemory(
+            CMem(inst.cbuf_slot, RZ, static_cast<u64>(inst.cbuf_offset * 4)),
+            ir::VectorType(ir::ScalarType::F16, 2)),
         inst.neg_c);
 }
 
@@ -206,8 +209,9 @@ void EmitHfma2C(DecoderContext& context, InstHfma2C inst) {
         context, inst.base.pred, inst.base.pred_inv, inst.sat,
         inst.base.out_fmt, inst.base.dst, inst.base.src_a, inst.base.swizzle_a,
         false,
-        ir::Value::ConstMemory(CMem(inst.cbuf_slot, RZ, inst.cbuf_offset * 4),
-                               ir::VectorType(ir::ScalarType::F16, 2)),
+        ir::Value::ConstMemory(
+            CMem(inst.cbuf_slot, RZ, static_cast<u64>(inst.cbuf_offset * 4)),
+            ir::VectorType(ir::ScalarType::F16, 2)),
         inst.neg_b,
         GetSwizzledHalf(context.builder, inst.swizzle_c, inst.src_c),
         inst.neg_c);
